+++ client/src/router/cognitoConfig.ts
@@ -4,6 +4,7 @@ interface OidcConfig {
   redirect_uri: string;
   scope: string;
   response_type: string;
+
   post_logout_redirect_uri: string;
 };
 
@@ -15,7 +16,7 @@ export const LOCAL_COGNITO_CONFIG: CognitoConfig = {
   authority: "https://cognito-idp.us-east-1.amazonaws.com/us-east-1_A7CaR2Wo3",
   domain: "https://us-east-1a7car2wo3.auth.us-east-1.amazoncognito.com",
   client_id: "5km9thunj8g6qd32s5et2i8pga",
-  post_logout_redirect_uri: "http://localhost:3000",
+  post_logout_redirect_uri: "http://localhost:3000/logout",
   redirect_uri: "http://localhost:3000/login-redirect",
   response_type: "code",
   scope: "openid email phone",
@@ -25,7 +26,7 @@ export const getCognitoLogoutUrl = (cognitoConfig: CognitoConfig): string => {
   const url = `${cognitoConfig.domain}/logout
   ?client_id=${cognitoConfig.client_id}
   &logout_uri=${encodeURIComponent(cognitoConfig.post_logout_redirect_uri)}`;
-  return url.replace(/\n/g, "");
+  return url.replace(/\s+/g, "");
 };
 
 export const logout = () => {
